+++ src/libWrapper-wrapper.js
@@ -583,17 +583,18 @@ export class Wrapper {
 	}
 
 	warn_classic_wrapper() {
-		let module_name = get_current_module_name();
-		module_name = module_name ? `\u00AB${module_name}\u00BB` : '\u00ABunknown\u00BB';
-
+		const module_name = get_current_module_name() ?? '\u00ABunknown\u00BB';
 		const affectedModules = this.get_affected_modules();
 
 		if(affectedModules.length > 0) {
-			LibWrapperStats.register_conflict(module_name, affectedModules, this.name);
-			LibWrapperNotifications.conflict(module_name, affectedModules, true, `Detected non-libWrapper wrapping of '${this.name}' by ${module_name}. This will potentially lead to conflicts.`);
+			const notify = LibWrapperStats.register_conflict(module_name, affectedModules, this.name);
 
-			if(DEBUG && console.trace)
-				console.trace();
+			if(notify) {
+				LibWrapperNotifications.conflict(module_name, affectedModules, true, `Detected non-libWrapper wrapping of '${this.name}' by '${module_name}'. This will potentially lead to conflicts.`);
+
+				if(DEBUG && console.trace)
+					console.trace();
+			}
 		}
 
 		if(!this.detected_classic_wrapper)